修改:   1.增加移动至.   2.增加问题管理内附件另存为...   3.增加问题管理内的高级查询中的问题创建人与问题解决人   4.任务单的计划天数可以设置小数位.
作者:龙仕云 2008-7-31

git-svn-id: https://devemanage.googlecode.com/svn/trunk@90 9ff6b1e3-de3f-0410-996c-bd7989423481

+++ database/SQLScript/LSD_CMM_TABLE.sql
@@ -207,10 +207,10 @@ create table TB_TASK(
 	ZTESTCASE       text,                                   /*测试用例*/
 	ZSTATUS         int not null,                           /*状态 待分发=0 ; 执行中=1 ; 撤消=2; 完成=3 ; 关闭=4;激活=5*/
 	ZDATE           datetime,                               /*制单时间*/
-	ZPALNDAY        int not null default 1,                  /*计划工期(天)*/
+	ZPALNDAY        float not null default 1,                  /*计划工期(天)*/
 
 	ZBEGINDATE      datetime,	                            /*任务开始时间 由任务执行人生成,这时状态变更为执行中*/
-	ZDAY            int,                                    /*实际的天数*/
+	ZDAY            float,                                    /*实际的天数*/
 	ZSUCCESSDATE    datetime,                               /*完成时间*/
 	ZCLOSEDATE      datetime,                               /*关闭时间*/
